Pulled out history tests into its own file.
