+++ test/runner
@@ -86,6 +86,7 @@ var files = [
   //"level3/ls.js",
   "level3/xpath.js",
   "window/index.js",
+  "window/history.js",
   "window/script.js",
   "window/console.js",
   "window/frame.js",
